extract main content of <App> into new component, <Main>
<Main> will sit within <App> and contain BottomNavigation, which has the 3 tabs LabelTab, MetricsTab, and ProfileSettings.
This will allow the tabs to share some information, but be categorically separate from OnboardingStack.
